+++ src/main/java/com/dku/council/domain/user/controller/UserController.java
@@ -8,6 +8,7 @@ import com.dku.council.domain.user.model.dto.response.ResponseLoginDto;
 import com.dku.council.domain.user.model.dto.response.ResponseMajorDto;
 import com.dku.council.domain.user.model.dto.response.ResponseRefreshTokenDto;
 import com.dku.council.domain.user.model.dto.response.ResponseUserInfoDto;
+import com.dku.council.domain.user.service.SignupService;
 import com.dku.council.domain.user.service.UserService;
 import com.dku.council.global.auth.jwt.AppAuthentication;
 import com.dku.council.global.auth.role.UserOnly;
@@ -29,7 +30,8 @@ import java.util.stream.Collectors;
 public class UserController {
 
     private final MessageSource messageSource;
-    private final UserService service;
+    private final UserService userService;
+    private final SignupService signupService;
 
 
     /**
@@ -40,7 +42,7 @@ public class UserController {
     @GetMapping
     @UserOnly
     public ResponseUserInfoDto getMyInfo(AppAuthentication auth) {
-        return service.getUserInfo(auth.getUserId());
+        return userService.getUserInfo(auth.getUserId());
     }
 
     /**
@@ -53,7 +55,7 @@ public class UserController {
     @PostMapping("/{signup-token}")
     public void signup(@Valid @RequestBody RequestSignupDto dto,
                        @PathVariable("signup-token") String signupToken) {
-        service.signup(dto, signupToken);
+        signupService.signup(dto, signupToken);
     }
 
     /**
@@ -64,7 +66,7 @@ public class UserController {
      */
     @PostMapping("/login")
     public ResponseLoginDto login(@Valid @RequestBody RequestLoginDto dto) {
-        return service.login(dto);
+        return userService.login(dto);
     }
 
     /**
@@ -86,7 +88,7 @@ public class UserController {
     @UserOnly
     public ResponseRefreshTokenDto refreshToken(HttpServletRequest request,
                                                 @Valid @RequestBody RequestRefreshTokenDto dto) {
-        return service.refreshToken(request, dto.getRefreshToken());
+        return userService.refreshToken(request, dto.getRefreshToken());
     }
 
     /**
