+++ src/main/java/com/dku/council/domain/user/service/UserService.java
@@ -1,22 +1,16 @@
 package com.dku.council.domain.user.service;
 
-import com.dku.council.domain.user.exception.AlreadyStudentIdException;
 import com.dku.council.domain.user.exception.LoginUserNotFoundException;
 import com.dku.council.domain.user.exception.WrongPasswordException;
-import com.dku.council.domain.user.model.MajorData;
-import com.dku.council.domain.user.model.UserStatus;
 import com.dku.council.domain.user.model.dto.request.RequestLoginDto;
-import com.dku.council.domain.user.model.dto.request.RequestSignupDto;
 import com.dku.council.domain.user.model.dto.response.ResponseLoginDto;
 import com.dku.council.domain.user.model.dto.response.ResponseRefreshTokenDto;
 import com.dku.council.domain.user.model.dto.response.ResponseUserInfoDto;
-import com.dku.council.domain.user.model.entity.Major;
 import com.dku.council.domain.user.model.entity.User;
 import com.dku.council.domain.user.repository.UserRepository;
 import com.dku.council.global.auth.jwt.AuthenticationToken;
 import com.dku.council.global.auth.jwt.JwtProvider;
 import com.dku.council.global.auth.role.UserRole;
-import com.dku.council.infra.dku.model.StudentInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.MessageSource;
@@ -25,10 +19,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.Optional;
 
 // TODO Test it
-// TODO Signup Service로 분리하면 좋을 것 같다.
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
@@ -36,52 +28,10 @@ import java.util.Optional;
 public class UserService {
 
     private final MessageSource messageSource;
-    private final SMSVerificationService smsVerificationService;
-    private final DKUAuthService dkuAuthService;
     private final PasswordEncoder passwordEncoder;
     private final UserRepository userRepository;
     private final JwtProvider jwtProvider;
 
-
-    @Transactional
-    public void signup(RequestSignupDto dto, String signupToken) {
-        StudentInfo studentInfo = dkuAuthService.getStudentInfo(signupToken);
-        String phone = smsVerificationService.getPhoneNumber(signupToken);
-        String encryptedPassword = passwordEncoder.encode(dto.getPassword());
-        Optional<User> alreadyUser = userRepository.findByStudentId(studentInfo.getStudentId());
-
-        if (alreadyUser.isPresent()) {
-            throw new AlreadyStudentIdException();
-        }
-
-        User.UserBuilder userBuilder = User.builder()
-                .studentId(studentInfo.getStudentId())
-                .password(encryptedPassword)
-                .name(studentInfo.getStudentName())
-                .phone(phone)
-                .yearOfAdmission(studentInfo.getYearOfAdmission())
-                .status(UserStatus.ACTIVE)
-                .role(UserRole.USER);
-
-        MajorData majorData = studentInfo.getMajorData();
-        if (majorData.isEmpty()) {
-            userBuilder.major(new Major(studentInfo.getNotRecognizedMajor(), studentInfo.getNotRecognizedDepartment()));
-        } else {
-            userBuilder.major(new Major(majorData));
-        }
-
-        userRepository.save(userBuilder.build());
-        deleteSignupAuths(signupToken);
-    }
-
-    private void deleteSignupAuths(String signupToken) {
-        if (!dkuAuthService.deleteStudentAuth(signupToken)) {
-            log.error("Can't delete user signup authentication: dku student auth");
-        }
-        if (!smsVerificationService.deleteSMSAuth(signupToken)) {
-            log.error("Can't delete user signup authentication: sms auth");
-        }
-    }
 
     public ResponseLoginDto login(RequestLoginDto dto) {
         User user = userRepository.findByStudentId(dto.getStudentId())
